requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-core
langchain-community
langchain-groq
langchain-chroma
langchain-huggingface
pdfminer.six
python-dotenv
pandas
pytesseract
opencv-python
sentence-transformers
transformers==4.40.2
torch
accelerate
python-docx
docx2txt
pillow
pdf2image
fpdf
PyPDF2
tiktoken
faiss-cpu
fastapi
uvicorn
langsmith

protobuf==4.25.3
tensorflow-intel==2.16.1
opentelemetry-proto==1.18.0
